refactor: Restructure framework package
