+++ src/main/java/com/tistory/framework/security/config/SecurityConfig.java
@@ -1,7 +1,7 @@
-package com.tistory.framework.config;
+package com.tistory.framework.security.config;
 
-import com.tistory.framework.filter.JwtRequestFilter;
-import com.tistory.framework.service.CustomUserDetailsService;
+import com.tistory.framework.security.filter.JwtRequestFilter;
+import com.tistory.framework.security.service.CustomUserDetailsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -13,8 +13,6 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-import static org.springframework.security.config.Customizer.withDefaults;
 
 @Configuration // Spring 설정 클래스임
 @EnableWebSecurity // Spring Security를 활성화
